refactor: restructure to layered architecture
- Extract shared theme to ui/core/theme.dart (remove 3x duplication)
- Restructure lib/ to data/ui/features/ui/shared layers
- Merge duplicate PostCard into single shared component with CachedNetworkImage
- Fix naming: Purrfect -> PurrCat
- Remove unused imports, fix SDK version mismatch
