+++ mobile/purrcat_app_mobile/lib/main.dart
@@ -1,22 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:go_router/go_router.dart';
 
-// Utils
 import 'utils/routes.dart';
-
-// Providers
-import 'providers/auth_provider.dart';
+import 'ui/features/auth/view_models/auth_provider.dart';
+import 'ui/core/theme.dart';
 
 void main() {
   runApp(const PurrCatApp());
 }
-
-// Global Theme Colors (corrected to match Figma design)
-const Color brandPink = Color(0xFFA03A57);
-const Color headingColor = Color(0xFF1A1A1A);
-const Color bodyColor = Color(0xFF757575);
-const Color backgroundColor = Colors.white;
 
 class PurrCatApp extends StatelessWidget {
   const PurrCatApp({super.key});
