+++ mobile/purrcat_app_mobile/lib/ui/core/theme.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/material.dart';
+
+const Color brandPink = Color(0xFFA03A57);
+const Color headingColor = Color(0xFF1A1A1A);
+const Color bodyColor = Color(0xFF757575);
+const Color backgroundColor = Colors.white;
