+++ mobile/purrcat_app_mobile/lib/ui/features/feed/views/feed_screen.dart
@@ -1,16 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:dotted_border/dotted_border.dart';
-import 'package:shimmer/shimmer.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 
-import '../models/feed_model.dart';
+import '../../../../data/models/feed_model.dart';
+import '../../../../ui/core/theme.dart';
+import '../../../../ui/shared/post_card.dart';
 
-// Global Theme Colors (corrected to match Figma design)
-const Color brandPink = Color(0xFFA03A57);
-const Color headingColor = Color(0xFF1A1A1A);
-const Color bodyColor = Color(0xFF757575);
-const Color backgroundColor = Colors.white;
+
 
 class FeedScreen extends StatefulWidget {
   const FeedScreen({super.key});
@@ -20,8 +17,6 @@ class FeedScreen extends StatefulWidget {
 }
 
 class _FeedScreenState extends State<FeedScreen> {
-  bool _isLiked = false;
-
   final List<String> _stories = [
     'https://i.pravatar.cc/150?img=1',
     'https://i.pravatar.cc/150?img=2',
@@ -83,7 +78,7 @@ class _FeedScreenState extends State<FeedScreen> {
             backgroundColor: backgroundColor,
             elevation: 0,
             title: Text(
-              'Purrfect',
+              'PurrCat',
               style: GoogleFonts.poppins(
                 fontSize: 20,
                 fontWeight: FontWeight.bold,
@@ -229,143 +224,3 @@ class _FeedScreenState extends State<FeedScreen> {
   }
 }
 
-// Post Card Widget
-class PostCard extends StatelessWidget {
-  final Post post;
-  final VoidCallback onLike;
-
-  const PostCard({
-    super.key,
-    required this.post,
-    required this.onLike,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: const EdgeInsets.only(bottom: 16),
-      padding: const EdgeInsets.symmetric(horizontal: 16),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          // User Header
-          ListTile(
-            contentPadding: EdgeInsets.zero,
-            leading: CircleAvatar(
-              radius: 20,
-              backgroundImage: CachedNetworkImageProvider(post.userAvatar),
-            ),
-            title: Text(
-              post.userName,
-              style: GoogleFonts.poppins(
-                fontSize: 14,
-                fontWeight: FontWeight.bold,
-                color: headingColor,
-              ),
-            ),
-            subtitle: Text(
-              'Location',
-              style: GoogleFonts.poppins(
-                fontSize: 12,
-                color: bodyColor,
-              ),
-            ),
-            trailing: const Icon(Icons.more_horiz, color: headingColor),
-          ),
-
-          // Media Card
-          Container(
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(24),
-              color: Colors.grey[200],
-            ),
-            child: ClipRRect(
-              borderRadius: BorderRadius.circular(24),
-              child: AspectRatio(
-                aspectRatio: 4 / 5,
-                child: CachedNetworkImage(
-                  imageUrl: post.images.isNotEmpty ? post.images[0] : '',
-                  fit: BoxFit.cover,
-                  placeholder: (context, url) => Shimmer.fromColors(
-                    baseColor: Colors.grey[300]!,
-                    highlightColor: Colors.grey[100]!,
-                    child: Container(
-                      color: Colors.grey[300],
-                    ),
-                  ),
-                  errorWidget: (context, url, error) => Container(
-                    color: Colors.grey[300],
-                    child: const Icon(Icons.error),
-                  ),
-                ),
-              ),
-            ),
-          ),
-
-          // Interaction Bar
-          Padding(
-            padding: const EdgeInsets.symmetric(vertical: 12),
-            child: Row(
-              children: [
-                IconButton(
-                  icon: Icon(
-                    post.isLiked ? Icons.favorite : Icons.favorite_border,
-                    color: post.isLiked ? brandPink : headingColor,
-                  ),
-                  onPressed: onLike,
-                ),
-                IconButton(
-                  icon: const Icon(Icons.chat_bubble_outline, color: headingColor),
-                  onPressed: () {},
-                ),
-                IconButton(
-                  icon: const Icon(Icons.send_rounded, color: headingColor),
-                  onPressed: () {},
-                ),
-                const Spacer(),
-                IconButton(
-                  icon: const Icon(Icons.bookmark_border, color: headingColor),
-                  onPressed: () {},
-                ),
-              ],
-            ),
-          ),
-
-          // Content
-          Padding(
-            padding: const EdgeInsets.only(bottom: 16),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  '${post.likes} Likes',
-                  style: GoogleFonts.poppins(
-                    fontSize: 14,
-                    fontWeight: FontWeight.bold,
-                    color: headingColor,
-                  ),
-                ),
-                const SizedBox(height: 4),
-                RichText(
-                  text: TextSpan(
-                    style: GoogleFonts.poppins(
-                      fontSize: 14,
-                      color: headingColor,
-                    ),
-                    children: [
-                      TextSpan(
-                        text: '${post.userName} ',
-                        style: const TextStyle(fontWeight: FontWeight.bold),
-                      ),
-                      TextSpan(text: post.content),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
