+++ mobile/purrcat_app_mobile/lib/ui/shared/bottom_nav_component.dart
@@ -1,10 +1,6 @@
 import 'package:flutter/material.dart';
 
-// Global Theme Colors (corrected to match Figma design)
-const Color brandPink = Color(0xFFA03A57);
-const Color headingColor = Color(0xFF1A1A1A);
-const Color bodyColor = Color(0xFF757575);
-const Color backgroundColor = Colors.white;
+import '../core/theme.dart';
 
 class BottomNavComponent extends StatelessWidget {
   final int currentIndex;
