+++ mobile/purrcat_app_mobile/lib/ui/shared/post_card.dart
@@ -1,0 +1,147 @@
+import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'package:shimmer/shimmer.dart';
+import 'package:cached_network_image/cached_network_image.dart';
+
+import '../../data/models/feed_model.dart';
+import '../core/theme.dart';
+
+class PostCard extends StatelessWidget {
+  final Post post;
+  final VoidCallback onLike;
+
+  const PostCard({
+    super.key,
+    required this.post,
+    required this.onLike,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.only(bottom: 16),
+      padding: const EdgeInsets.symmetric(horizontal: 16),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          // User Header
+          ListTile(
+            contentPadding: EdgeInsets.zero,
+            leading: CircleAvatar(
+              radius: 20,
+              backgroundImage: CachedNetworkImageProvider(post.userAvatar),
+            ),
+            title: Text(
+              post.userName,
+              style: GoogleFonts.poppins(
+                fontSize: 14,
+                fontWeight: FontWeight.bold,
+                color: headingColor,
+              ),
+            ),
+            subtitle: Text(
+              'Location',
+              style: GoogleFonts.poppins(
+                fontSize: 12,
+                color: bodyColor,
+              ),
+            ),
+            trailing: const Icon(Icons.more_horiz, color: headingColor),
+          ),
+
+          // Media Card
+          Container(
+            decoration: BoxDecoration(
+              borderRadius: BorderRadius.circular(24),
+              color: Colors.grey[200],
+            ),
+            child: ClipRRect(
+              borderRadius: BorderRadius.circular(24),
+              child: AspectRatio(
+                aspectRatio: 4 / 5,
+                child: CachedNetworkImage(
+                  imageUrl: post.images.isNotEmpty ? post.images[0] : '',
+                  fit: BoxFit.cover,
+                  placeholder: (context, url) => Shimmer.fromColors(
+                    baseColor: Colors.grey[300]!,
+                    highlightColor: Colors.grey[100]!,
+                    child: Container(
+                      color: Colors.grey[300],
+                    ),
+                  ),
+                  errorWidget: (context, url, error) => Container(
+                    color: Colors.grey[300],
+                    child: const Icon(Icons.error),
+                  ),
+                ),
+              ),
+            ),
+          ),
+
+          // Interaction Bar
+          Padding(
+            padding: const EdgeInsets.symmetric(vertical: 12),
+            child: Row(
+              children: [
+                IconButton(
+                  icon: Icon(
+                    post.isLiked ? Icons.favorite : Icons.favorite_border,
+                    color: post.isLiked ? brandPink : headingColor,
+                  ),
+                  onPressed: onLike,
+                ),
+                IconButton(
+                  icon: const Icon(Icons.chat_bubble_outline, color: headingColor),
+                  onPressed: () {},
+                ),
+                IconButton(
+                  icon: const Icon(Icons.send_rounded, color: headingColor),
+                  onPressed: () {},
+                ),
+                const Spacer(),
+                IconButton(
+                  icon: const Icon(Icons.bookmark_border, color: headingColor),
+                  onPressed: () {},
+                ),
+              ],
+            ),
+          ),
+
+          // Content
+          Padding(
+            padding: const EdgeInsets.only(bottom: 16),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  '${post.likes} Likes',
+                  style: GoogleFonts.poppins(
+                    fontSize: 14,
+                    fontWeight: FontWeight.bold,
+                    color: headingColor,
+                  ),
+                ),
+                const SizedBox(height: 4),
+                RichText(
+                  text: TextSpan(
+                    style: GoogleFonts.poppins(
+                      fontSize: 14,
+                      color: headingColor,
+                    ),
+                    children: [
+                      TextSpan(
+                        text: '${post.userName} ',
+                        style: const TextStyle(fontWeight: FontWeight.bold),
+                      ),
+                      TextSpan(text: post.content),
+                    ],
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
